requirements.txt
flask>=3.0.0
pillow>=10.0.0
flask-limiter>=3.5.0
gunicorn>=21.0.0
